Data collection job trigger complete with db changes.

+++ mipa-parent/mipa-osgi-modules/mipa-osgi-core/src/main/java/com/mipa/osgi/core/db/entity/ProviderDataEntity.java
@@ -7,7 +7,7 @@ public class ProviderDataEntity {
 	
 	private String providerName;
 	
-	private Short countryId;
+	private Integer countryId;
 	
 	private String pincodes;
 	
@@ -21,11 +21,11 @@ public class ProviderDataEntity {
 		this.providerId = providerId;
 	}
 
-	public Short getCountryId() {
+	public Integer getCountryId() {
 		return countryId;
 	}
 
-	public void setCountryId(Short countryId) {
+	public void setCountryId(Integer countryId) {
 		this.countryId = countryId;
 	}
